Extract SectionHighlights component
