+++ src/app/content/highlights/components/Highlights.spec.tsx
@@ -22,12 +22,13 @@ import {
 import { highlightLocationFilters } from '../selectors';
 import { SummaryHighlights } from '../types';
 import { getHighlightLocationFilterForPage } from '../utils';
-import Highlights, { SectionHighlights } from './Highlights';
+import Highlights from './Highlights';
 import { HighlightSection, LoaderWrapper } from './ShowMyHighlightsStyles';
 import ContextMenu from './SummaryPopup/ContextMenu';
 import HighlightAnnotation from './SummaryPopup/HighlightAnnotation';
 import HighlightDeleteWrapper from './SummaryPopup/HighlightDeleteWrapper';
 import { HighlightContentWrapper } from './SummaryPopup/HighlightListElement';
+import { SectionHighlights } from './SummaryPopup/SectionHighlights';
 
 const hlBlue = { id: 'hl1', color: HighlightColorEnum.Blue, annotation: 'hl1' };
 const hlGreen = { id: 'hl2', color: HighlightColorEnum.Green, annotation: 'hl' };
